refactor(CollectionCard): Extract title rendering logic into separate components
