+++ packages/components/src/templates/next/components/internal/CollectionCard/Title/index.ts
@@ -1,0 +1,1 @@
+export * from "./Title"
